S0C-3A-1A: Implement shadow_verify_canary_dual_write scenario and update related CLI and documentation
